fixed correspond_idx_ bug in sfm and add pose pnp test

+++ test/map_test.cpp
@@ -8,7 +8,7 @@
 #include <fstream>
 using namespace std;
 
-void savePoints(const string &name,std::vector<cv::Point3d> points){
+void savePoints(const string &name,std::vector<cv::Point3f> points){
 	ofstream out(name);
 	for(auto p: points){
 		out << p.x << "," << p.y << "," << p.z << endl;
